fix(ui): unblock environment picker CI checks
